fix(ui): update folder naming and MinimizedFileCard menu usage
