+++ app/shared/components/cards-layout/CardMenu.styles.ts
@@ -1,11 +1,6 @@
 import { SxProps, Theme } from '@mui/material';
 
 const styles: Record<string, SxProps<Theme>> = {
-  menu: {
-    '& .MuiPaper-root': {
-      padding: '0px'
-    }
-  },
   menuItem: {
     padding: '8px 16px',
     '&:hover': {
